backend/requirements.txt
django
youtube-transcript-api
requests
python-dotenv
gunicorn
webvtt-py
yt-dlp